Split lib.common into one file per class
* Create one file per class in lib/
* Move XO-related functions to lib/xo.py
* Temporarily solve import cycle (lib.pool imports lib.host and vice versa) with
  an in-function import. A better fix follows in a separate commit.
