+++ lib/vif.py
@@ -1,0 +1,28 @@
+import lib.commands as commands
+
+class VIF:
+    def __init__(self, uuid, vm):
+        self.uuid = uuid
+        self.vm = vm
+
+    def param_get(self, param_name, key=None, accept_unknown_key=False):
+        args = {'uuid': self.uuid, 'param-name': param_name}
+        if key is not None:
+            args['param-key'] = key
+        try:
+            value = self.vm.host.xe('vif-param-get', args)
+        except commands.SSHCommandFailed as e:
+            if key and accept_unknown_key and e.stdout == "Error: Key %s not found in map" % key:
+                value = None
+            else:
+                raise
+        return value
+
+    def device_id(self):
+        """ Build the identifier that will allow to grep for the VIF's interrupts. """
+        dom_id = self.vm.param_get('dom-id')
+        device = self.param_get('device')
+        return f"vif{dom_id}.{device}"
+
+    def move(self, network_uuid):
+        self.vm.host.xe('vif-move', {'uuid': self.uuid, 'network-uuid': network_uuid})
