+++ lib/xo.py
@@ -1,0 +1,18 @@
+import json
+import subprocess
+
+def xo_cli(action, args={}, check=True, simple_output=True):
+    res = subprocess.run(
+        ['xo-cli', action] + ["%s=%s" % (key, value) for key, value in args.items()],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+        check=check
+    )
+    if simple_output:
+        return res.stdout.decode().strip()
+    else:
+        return res
+
+def xo_object_exists(uuid):
+    lst = json.loads(xo_cli('--list-objects', {'uuid': uuid}))
+    return len(lst) > 0
